// src/types/delivery.ts
export interface IDeliveryStore {
    controller: null | AbortController;
    addressObject: IAddressObject | null;
    amount: number;
    isLoading: boolean;
}
export type IAddressObject = {
    value: string;
    unrestricted_value: string;
    data: {
        postal_code: string;
        country_iso_code: string;
        country: string;
        federal_district: string;
        region_fias_id: string;
        region_kladr_id: string;
        region_iso_code: string;
        region_with_type: string;
        region_type: string;
        region_type_full: string;
        region: string;
        area_fias_id: string;
        area_kladr_id: string;
        area_type: string;
        area_type_full: string;
        area: string;
        sub_area_fias_id: string;
        sub_area_kladr_id: string;
        sub_area_with_type: string;
        sub_area_type: string;
        sub_area: string;
        city: string;
        city_fias_id: string;
        city_kladr_id: string;
        city_with_type: string;
        city_type: string;
        city_type_full: string;
        city_district_fias_id: string;
        city_district_kladr_id: string;
        city_district_with_type: string;
        city_district_type: string;
        city_district_type_full: string;
        settlement_fias_id: string;
        settlement_kladr_id: string;
        settlement_with_type: string;
        settlement_type: string;
        settlement_type_full: string;
        settlement: string;
        street_fias_id: string;
        street_kladr_id: string;
        street_type: string;
        street_type_full: string;
        street: string;
        stead_fias_id: string;
        stead_type: string;
        stead: string;
        house: string;
        house_fias_id: string;
        house_kladr_id: string;
        house_type: string;
        house_type_full: string;
        block_type: string;
        block_type_full: string;
        block: string;
        flat: string;
        flat_fias_id: string;
        flat_type: string;
        flat_type_full: string;
        room_fias_id: string;
        room_type: string;
        room_type_full: string;
        room: string;
        postal_box: string;
        fias_id: string;
        kladr_id: string;
        geoname_id: string;
        capital_marker: string;
        okato: string;
        oktmo: string;
        tax_office: string;
        source: string;
        history_values: string[];
        ///
        geo_lat?: string;
        geo_lon?: string;
        qc_geo?: string;
        fias_actuality_state?: string;
        city_area?: string;
        //
        beltway_hit?: string;
        beltway_distance?: string;
        stead_cadnum?: string;
        house_cadnum?: string;
        flat_cadnum?: string;
        flat_area?: string;
        square_meter_price?: string;
        room_cadnum?: string;
        timezone?: string;
        metro?: {
            name: string;
            line: string;
            distance: number;
        }[];
    };
};